feature: se añade seguridad basada en roles y JWT.

+++ domain/model/src/main/java/co/com/zenway/model/rol/Rol.java
@@ -9,7 +9,7 @@ import lombok.*;
 @Builder(toBuilder = true)
 public class Rol {
 
-    private Long id;
+    private Short id;
     private String nombre;
     private String descripcion;
